implementing attaching files on modify document

+++ src/components/board/editor/Editor.vue
@@ -36,8 +36,9 @@
         </v-btn>
       </div>
       <div>
-        <v-btn small flat @click="$refs.pond.browse()">
+        <v-btn small flat @click="attachButtonClick">
           <v-icon id="attach-button">attach_file</v-icon>파일첨부
+          <input type="file" multiple id="file-upload" style="display:none" @change="onFileChange">
         </v-btn>
       </div>
     </v-layout>
@@ -48,7 +49,7 @@
             <v-layout row align-center>
               <div class="ellipsis">{{item}}</div>
               <v-spacer/>
-              <v-btn small class="short" @click="removeFile(item)">삭제</v-btn>           
+              <v-btn small class="short" @click="removeFile(item)">삭제</v-btn>
             </v-layout>
           </v-slide-y-transition>
         </v-flex>
@@ -139,9 +140,12 @@ export default {
       isAnonymous: false,
       disallowAnonymous: false,
       attachedFilenames: [],
+      deletedFilenames: [],
       attachedImages: [],
       formData: undefined,
+      modifyingFormData: undefined,
       attachedFileNumber: 0,
+      attachFromServer: undefined,
       selectedFile: undefined,
       originImages: [],
       server: {
@@ -216,7 +220,8 @@ export default {
     },
     async modifyPost() {
       // edit given document put 한번 attach post delete한번
-      if (!this.formData) this.formData = new FormData();
+      // if (!this.formData) this.formData = new FormData();
+
       // check image and attach change
       // 이미지/파일 달라진 것 -> post/delete
       // 이전 파일 올릴 때 동일이름 체크 필요(이미지는 uid부여중)
@@ -226,12 +231,21 @@ export default {
         title: this.title,
         contents: JSON.stringify(this.$refs.editor.quill.editor.delta)
       }
+      // let fileDeleteList = []
+      // let fileNewList = []
+      console.log(this.attachFromServer)
+      this.deletedFilenames.forEach(f => {
+        console.log(f)
+        
+      })
       return this.$axios
       .put(`/document`, modifiedBody)
       .then(response => {
         if (response.status === 200) {
           console.log(response)
-          this.$router.push(`/${this.$route.params.boardId}/${this.documentId}`);
+
+          console.log(this.deletedFilenames)
+          // this.$router.push(`/${this.$route.params.boardId}/${this.documentId}`);
         }
       })
       .catch(error => {
@@ -261,9 +275,21 @@ export default {
       }
     },
     removeFile(filename) {
-      let fileid = this.$refs.pond.getFiles().find(file => file.filename === filename).id;
-      this.$refs.pond.removeFile(fileid);
-      this.show = false;
+      if(this.documentId) {
+        // when editing
+        this.deletedFilenames.push(filename)
+        this.attachedFilenames = this.attachedFilenames.filter(f => f !== filename)
+        // 없는 애들로 새로 생성해서 올려야 함.
+        // Display the key/value pairs
+        // for (var pair of this.modifyingFormData.entries()) {
+        //     console.log(pair[0]+ ', ' + pair[1]); 
+        //     console.log(pair[1].name)
+        // }
+      } else {
+        let fileid = this.$refs.pond.getFiles().find(file => file.filename === filename).id;
+        this.$refs.pond.removeFile(fileid);
+        this.show = false;
+      }
     },
     handleFilePondRemoveFile: function(file) {
       this.attachedFilenames = this.attachedFilenames.filter(filename => file.filename !== filename);
@@ -314,6 +340,10 @@ export default {
       }
       this.surveyDialog = true;
     },
+    attachButtonClick() {
+      if(this.documentId) this.openFileDialog()
+      else this.$refs.pond.browse()
+    },
     extractSurvey(survey) {
       this.surveyDialog = false;
       this.survey = JSON.parse(JSON.stringify(survey));
@@ -332,12 +362,42 @@ export default {
           this.boardId = data.boardId
           this.title = data.title
           this.contents = JSON.parse(data.contents)
+          this.attachFromServer = data.attach
           this.$refs.editor.quill.setContents(this.contents)
           this.isAnonymous = data.isAnonymous
           if(data.survey) {
             this.survey = data.survey
           }
-    }
+          if(data.attach) {
+            console.log(data.attach)
+            console.log(this.contents)
+            let image;
+            this.contents.ops.forEach(item => {
+              if (item.insert.hasOwnProperty("image")) {
+                image = data.attach.find(x => x.attach_name === item.insert.image);
+                if (image) {
+                  image.insert = true;
+                }
+              }
+            })
+            this.attachedFilenames = data.attach.filter(f => !f.insert).map(f => f.attach_name)
+            
+          }
+    },
+    openFileDialog() {
+      document.getElementById('file-upload').click();
+    },
+    onFileChange(e) {
+        if (!this.modifyingFormData) this.modifyingFormData = new FormData();
+        var self = this;
+        var files = e.target.files || e.dataTransfer.files;       
+        if(files.length > 0){
+            for(var i = 0; i < files.length; i++){
+                self.modifyingFormData.append("file", files[i], files[i].name);
+                self.attachedFilenames.push(files[i].name)
+            }
+        }
+    },
   },
   computed: {
     editor() {
